// operator-task/src/types/variant-type.ts
export enum VariantType {
  Uninitialized = 0,
  EncodeAnytype = 4,
  SInt32 = 327680,
  SInt64 = 4587520,
  UInt8 = 655360,
  UInt32 = 393216,
  Bit32 = 589824,
  Bit64 = 4980736,
  UInt64 = 4784128,
  Double = 458752,
  Bool = 262144,
  DateTime = 196608,
  StringType = 524288, // String in .net
  DpIdentifier = 720896,
  LangText = 2621440,
  Blob = 3014656,
  Anytype = 1572864,
  SInt32Array = 983040,
  SInt64Array = 4653056,
  UInt8Array = 1310720,
  UInt32Array = 1048576,
  UInt64Array = 4849664,
  Bit32Array = 1245184,
  Bit64Array = 5046272,
  DoubleArray = 1114112,
  BoolArray = 917504,
  TimeStampArray = 851968,
  StringArray = 1179648,
  DpIdentifierArray = 1376256,
  LangTextArray = 2686976,
  BlobArray = 3080192,
  VariantArray = 2293760,
  RecVar = 1441792
}